langchain-prompts/requirements.txt
# LangChain Core
langchain
langchain-core


langchain-openai
openai

langchain-google-genai
google-generativeai


python-dotenv
streamlit


